+++ example/lib/app/app_module.dart
@@ -1,16 +1,14 @@
-import 'package:example/app/home/home_store.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
 import 'modules/home/home_module.dart';
 
 class AppModule extends Module {
   @override
-  final List<Bind> binds = [
-    Bind.lazySingleton((i) => HomeStore()),
-  ];
+  final List<Bind> binds = [];
 
   @override
   final List<ModularRoute> routes = [
     ModuleRoute(Modular.initialRoute, module: HomeModule()),
   ];
+
 }